points logic done

+++ include/trainer_rank.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_TRAINER_RANK_H
 #define GUARD_TRAINER_RANK_H
 
-void AddTrainerPoints(u16 points);
+void AddTrainerPoints(u8 points);
 u8 GetTrainerRank(void);
 u16 GetTrainerPoints(void);
 u8 GetRankGoal(void);
